squash-to: "mptcp: track some aggregate data counters"
Similar to TCP, we can't acquire the socket lock if msk is in
TCP_LISTEN status or we risk deadlock.

Note that the state test is safe since mptcp_listen() moves the msk
socket into the listener status before acquiring the listener hash
locks and such test is under the listener hash losk scope, so we
always observe up-to-date status WRT listener sockets.

Signed-off-by: Paolo Abeni <pabeni@redhat.com>
Link: https://lore.kernel.org/r/829e91b37a2fc12a9a43f20204d26adc0572a7d3.1685963110.git.pabeni@redhat.com
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>