selftests: mptcp: connect: don't stop if error
No more tests were executed after a failure but it is still interesting
to get results for all the tests to better understand what's still OK
and what's not after a modification.

Now we only exit earlier if the basic tests are failing: no ping going
through namespaces or unable to transfer data on the loopback interface.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>
Message-Id: <20230621-mptcp-issue-368-selftests-subtests-v1-6-0cae453ae1fa@tessares.net>