net: socket: add socket_create wrapper
The arguments of socket() need to be changed sometimes like the MPTCP
case in the next commits.

It's too late to add the BPF hooks in BPF_CGROUP_RUN_PROG_INET_SOCK()
in inet_create(). So this patch adds a new wrapper socket_create().

The arguments of the wrapper need be changed through the BPF traces.
So add a new struct socket_args to put together all arguments, and pass
a pointer of this struct to the wrapper.

Signed-off-by: Geliang Tang <geliang.tang@suse.com>
Message-Id: <bf1cd56ede55306b3272c6b3069830ea6a189c97.1688366249.git.geliang.tang@suse.com>