bpf: Add bpf_mptcpify helper
This patch implements a new struct bpf_func_proto bpf_mptcpify_proto. And
define a new helper bpf_mptcpify() to mptcpify a TCP socket dynamically as
an MPTCP one when it is created.

In bpf_mptcpify(), if the protocol ID of sk is IPPROTO_TCP, set it to
IPPROTO_MPTCP.

Signed-off-by: Geliang Tang <geliang.tang@suse.com>
Message-Id: <d746717840dfdcab3e4d8fa710dfaa3958a0ead1.1688366249.git.geliang.tang@suse.com>