selftests: mptcp: pm_nl_ctl: always look for errors
If a Netlink command for the MPTCP path-managers is not valid, it is
important to check if there are errors. If yes, they need to be reported
instead of being ignored and exiting without errors.

Now if no replies are expected, an ACK from the kernelspace is asked by
the userspace in order to always expect a reply. We can use the same
buffer that is currently always >1024 bytes. Then we can check if there
is an error (err->error), print it if any and report the error.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>
Message-Id: <20230705-mptcp-issue-368-pm_nl_ctl_ack-v1-1-40bcff40cb6b@tessares.net>