Squash to "selftests: mptcp: join: colored results"
New env var to force using colours.

Most CIs these days support colours in the serial but they will execute
scripts in a non interactive mode and colours will then not be displayed
by default. So we need a way to force displaying them when we are in
this mode:

  SELFTESTS_MPTCP_LIB_COLOR_FORCE=1

Note that it is still possible to force disabling them if needed via the
standard NO_COLOR=1 env var way.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>
Message-Id: <20230728141833.916787-1-matthieu.baerts@tessares.net>