Squash to "selftests: mptcp: pm_nl_ctl: always look for errors"
mptcp_join.sh selftest tries to add a bad endpoint on purpose ("ID
change is prevented") so it is normal to see this error:

   114 implicit EP
         creation                            [ ok ]
         mptcp_info subflows=1:1             [ ok ]
         mptcp_info add_addr=1:1             [ ok ]
   netlink error -22 (Invalid argument)
   ./pm_nl_ctl: bailing out due to netlink error[s]
         ID change is prevented              [ ok ]
         modif is allowed                    [ ok ]

We can mute the output because we expect an error.

We also need to edit the commit message to mention mptcp_join.sh
selftest also need to be modified, not just pm_netlink.sh.

Link: https://lore.kernel.org/r/20230728152845.1020528-1-matthieu.baerts@tessares.net
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>