Revert "ci: reduce global timeout to 90m"
This reverts commit cb993082cd3cf492c64dc314ea7b0e4223d6f21a.

When we switch to a new major version, ccache might not be able to help
accelerating the build step. In that case (and because only 2 CPU cores
are being used), it can be very slow to build the kernel (e.g. ~1h)
leaving not enough time to do the tests with a debug kernel.

The CI scripts have been modified to have a dedicated and short timeout
in case of boot issue. In that case, we avoid situations where the VM
stalls for 1h before starting the tests. We can then put the 2h global
timeout back.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>