utils: script to apply patches for CI
Sometimes, patchew cannot apply some patches due to conflicts: different
base, dependences, etc.

This new script applies the patches with b4, then replace the Link with
the Message-Id that is required by the CI, then create a tag and push
it.

By doing that with a script, we avoid manual errors: missing the
Message-Id or wrongly formatted, etc.

Signed-off-by: Matthieu Baerts <matttbe@kernel.org>