mptcp: drop lookup_by_id parameter in lookup_addr
When the lookup_by_id parameter of __lookup_addr() is true, it's the same
as __lookup_addr_by_id(), it can be replaced by __lookup_addr_by_id()
directly. So drop this parameter, let __lookup_addr() only looks up address
on the local address list by comparing addresses in it, not address ids.

Signed-off-by: Geliang Tang <geliang.tang@suse.com>
Message-Id: <43520f8214c70ec442c5f9b9cb9e56b90b75ba31.1699843617.git.geliang.tang@suse.com>