mptcp: drop lookup_by_id parameter in lookup_addr
When the lookup_by_id parameter of __lookup_addr() is true, it's the same
as __lookup_addr_by_id(), it can be replaced by __lookup_addr_by_id()
directly. So drop this parameter, let __lookup_addr() only looks up address
on the local address list by comparing addresses in it, not address ids.

Signed-off-by: Geliang Tang <geliang.tang@suse.com>
Message-Id: <8c13ef7d113ae1b870ad2ef47d70a795afb12812.1700312872.git.geliang.tang@suse.com>