gh: notif: no notif if skipped
Currently, tests have been disabled because we reached the monthly
limit. But we still have the notifications.

We now simply send notifications if the conclusion is not "skipped".

e.g.:

  https://api.github.com/repos/multipath-tcp/mptcp_net-next/check-suites/19169762939

Note that the conclusion is also available in each "check-runs":

  https://api.github.com/repos/multipath-tcp/mptcp_net-next/check-suites/19169762939/check-runs?status=completed

I'm not sure we can easily detect that we reached the monthly limit by
just looking at the check-suite:

  https://api.github.com/repos/multipath-tcp/mptcp_net-next/check-suites/19167594579

We could look at each "check-runs" but well, we still need to send a
notification that we reached the limit (and best to switch to a
different service not having such small limit):

  https://api.github.com/repos/multipath-tcp/mptcp_net-next/check-suites/19167594579/check-runs?status=completed

Signed-off-by: Matthieu Baerts <matttbe@kernel.org>