cirrus: only use it with a debug kernel
Cirrus has KVM support, it has the minimal resources to executes the
tests.

But we don't have enough "free" resources to run both the debug and
non-debug kconfig on Cirrus CI. So here, we decided to use it only to
run the tests with the debug kconfig, something that doesn't work on CI
not supporting KVM, e.g. on GH Action, it takes 6h and a lot of
instabilities.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>