gh: new workflow to run tests (non-debug)
Cirrus CI was working pretty well for us until a few months ago where
the free usage has been limited [1].

As a first step, we are now using GH Actions to run tests without a
debug kernel config.

We could use a GH Actions from Cirrus-Ci to ease the migration by using
their executor, but other adaptations are needed anyway. So here is a
new GH Actions executing the tests in the same docker. The notification
part has been duplicated and adapted. That should be temp if we decide
to switch to GH Actions only with custom resources.

This notif part has not been validated yet.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>