mptcp: bpf: fix compilation error without BPF_JIT
'bpf_mptcp_sched_ops' is only declared is CONFIG_BPF_JIT is set.

I guess we don't want to define it if we don't need it. We could add a
new helper doing nothing if BPF_JIT is not enabled, but it is probably
enough to just add a ifdef here, in the init function.

Fixes: 8ac5b9528258 ("mptcp: bpf: adapt code to match upstream changes")
Fixes: e76812a9ec49 ("bpf: Add bpf_mptcp_sched_ops")
Reported-by: kernel test robot <lkp@intel.com>
Closes: https://lore.kernel.org/oe-kbuild-all/202401312020.oG1MfzT1-lkp@intel.com/
Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>
Message-Id: <20240131134421.1045648-2-matttbe@kernel.org>