tcp: check the protocol with DEBUG_NET
Fuzzers and static checkers might not detect when tcp_sk() is used with
a non tcp_sock structure.

This kind of mistake already happened a few times with MPTCP, when
wrongly using TCP-specific helpers with mptcp_sock pointers. On the
other hand, there are many 'tcp_xxx()' helpers that are taking a 'struct
sock' as arguments, and some of them are only looking at fields from
'struct sock', and nothing from 'struct tcp_sock'. It is then tempting
to use them with a 'struct mptcp_sock'.

So a new simple check is done when CONFIG_DEBUG_NET is enabled. tcp_sk()
is then used as an inlined function, like before commit e9d9da91548b
("tcp: preserve const qualifier in tcp_sk()").

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>
Message-Id: <20240131-mptcp-check-protocol-v1-2-a06067f0bd08@kernel.org>