mptcp: check the protocol with DEBUG_NET
Fuzzers and static checkers might not detect when mptcp_sk() is used
with a non mptcp_sock structure.

This is similar to the parent commit, where it is easy to use mptcp_sk()
with a TCP sock, e.g. with a subflow sk.

So a new simple check is done when CONFIG_DEBUG_NET is enabled.
mptcp_sk() is then used as an inlined function, like before commit
403a40f2304d ("mptcp: preserve const qualifier in mptcp_sk()").

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>
Message-Id: <20240131-mptcp-check-protocol-v1-3-a06067f0bd08@kernel.org>