selftests: mptcp: connect: use += operator to append strings
This patch uses addition assignment operator (+=) to append strings in
mptcp_connect.sh.

For extra_args, use += instead of open-coding to make the statements
shorter.

Add a local variable extra in do_transfer to save the various extra warning
logs, using += to append it. And add a new variable tc_info to save various
tc info, also using += to append it. This can make the code more readable
and prepare for the next commit.

Signed-off-by: Geliang Tang <tanggeliang@kylinos.cn>
Message-Id: <bd6d215ffac737df7bc6c3d8780a4104810fe308.1709615883.git.tanggeliang@kylinos.cn>