gh: email: reply to the cover-letter
If available, and if we can reach PatchWork.

Instead of replying to the last message, with a different subject: some
email clients don't like that, and mark the email as a new thread. And
it is easier to use the cover-letter name, than the last patch, to know
about what this is.

We try to use the "full" subject if there is no cover-letter, e.g.

  [PATCH mptcp-next] mptcp: add X feature

Instead of

  mptcp: add X feature

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>