selftests: mptcp: iperf mptcp tests
MPTCPv1 protocol support has been added in iperf3 recently. This patch
add some tests for iperf3 to demonstrate the advantages of MPTCP over
TCP in latency or packet loss situations.

Note:

Does 'delay 1' in below command mean 'delay 1ms'? I'm not sure.
> tc -n $ns2 qdisc add dev ns2eth1 root netem rate 1mbit delay 1

Should this test be put in a standalone test script or just like in this
patch, as a subtest of mptcp_join.sh?

Signed-off-by: Geliang Tang <tanggeliang@kylinos.cn>
Message-Id: <8f95589ded2c77edbc5086a808703c63b4641e60.1710206580.git.tanggeliang@kylinos.cn>