gh: tests: publish failed results
Only the ones from the export and export-net branches, running on our
repo. This will let us the possibility to track the unstable tests.

The CI will then fetch the 'tests results' repo, then it will manipulate
the JSON:

 - append the tags (because we are executed from a branch to be able to
   save the cache)
 - merge the new results with the old ones
 - and remove old entries (>100)

Then the modifications will be pushed, and the new results will be
published by another Action.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>