gh: tests: exit with error on fork repos if unstable
If some tests are unstable, the workflow will end with a success to be
able to record the different results.

When this workflow is used on a fork, it is good to have GitHub sending
a notification in case of errors. Otherwise, when looking at the list of
Actions, we could think everything is OK because the Action ran without
(critical) error.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>