gh: patchew: workaround for empty push
When sending patches upstream, Patchew will try to apply them on top of
the export branch. 'git am' might not do anything, but it will end
with no errors. Patchew will then create a new tag at the top of the
export branch.

We can detect such issues, by looking for 'Message-Id:' in the commit
message, and skip some actions.

If we don't do that, notifications are sent on the mailing list for the
"DO-NOT-MERGE: mptcp: enabled by default" patch, and the jobs wait 30
minutes hoping to find this patch on Patchwork. So best to avoid that.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>