selftests/bpf: Add mptcp pm_nl_ctl link
This patch adds a link of mptcp pm_nl_ctl tool into bpf selftests, and
updates Makefile to compile it.

This is useful to run mptcp bpf selftests on a system with an old version
of IPRoite2 is installed. When 'ip mptcp' is not supported, this tool
pm_nl_ctl can be used instead.

Signed-off-by: Geliang Tang <tanggeliang@kylinos.cn>
Message-Id: <09e75e5622ae54bd6e160993ccf0a55b16139c71.1715431177.git.tanggeliang@kylinos.cn>