selftests/bpf: Add mptcp pm_nl_ctl link
This patch adds a link of mptcp pm_nl_ctl tool into bpf selftests, and
updates Makefile to compile it.

This is useful to run mptcp bpf selftests on a system with an old version
of IPRoite2 is installed. When 'ip mptcp' is not supported, this tool
pm_nl_ctl can be used instead.

Signed-off-by: Geliang Tang <tanggeliang@kylinos.cn>
Message-Id: <dcd8d80fab23f5c934ed7d4c6129cc517e1c504c.1715472312.git.tanggeliang@kylinos.cn>