selftests: mptcp: diag: trap cleanup after ns_init
mptcp_lib_ns_init may fail sometimes, then ns isn't setup correctly. In
that case, cleanup is invoked, and ns is used in cleanup. So an "Invalid
netns name ''" error occurs. The patch moves "trap cleanup" from the front
of mptcp_lib_ns_init to the back of it to fix this.

Signed-off-by: Geliang Tang <tanggeliang@kylinos.cn>
Message-Id: <f32ec0ed692ee29440c9507991e8f1903331d2b0.1716533107.git.tanggeliang@kylinos.cn>