selftests: net: lib: rename ns in setup_ns
The helpers setup_ns() doesn't work when a net namespace named "ns" is
passed to them.

For example, in net/mptcp/diag.sh, the name of the namespace is "ns".
If "setup_ns ns" is used in it, diag.sh fails with errors:

  Invalid netns name "./mptcp_connect"
  Cannot open network namespace "10000": No such file or directory
  Cannot open network namespace "10000": No such file or directory

That is because "ns" is also a local variable in setup_ns, and it will
not set the value for the global variable that has been giving in
argument. To solve this, this patch renames the local variable "ns" as
"_ns", which is more unlikely to conflict with existing variables. "_ns"
name appears to be unused in the net selftests.

Also a check has been added to setup_ns: if "_ns" is passed, setup_ns()
helper will exit with an error.

Signed-off-by: Geliang Tang <tanggeliang@kylinos.cn>
Message-Id: <20240524-selftests-net-lib-fixes-v5-7-b9e0968571a3@kernel.org>