selftests: mptcp: join: print title at the end of reset
init or init_partial may fail sometimes, in that case, titles shouldn't
be printed out.

For example, these errors occur when invalid name is set to setup_ns:

    001 no JOIN
    Failed to setup namespace 'ns_temp': invalid name

"001 no JOIN" shouldn't be printed out in this case. So the patch moves
mptcp_lib_print_title from the front of init to the end of reset().

Signed-off-by: Geliang Tang <tanggeliang@kylinos.cn>
Message-Id: <7dbbb909edef2fecc46bde435fae21be6bc9ed81.1716641976.git.tanggeliang@kylinos.cn>